moved helper function language_check() and yaml loader to separate file
